Python/FastApi/ai-learning-assistant/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
sqlalchemy
asyncpg
python-dotenv
openai
tortoise-orm[asyncpg]
